pyproject.toml
[project]
name = "voice-to-text"
version = "0.1.0"
description = "A CLI tool to convert m4a audio files to text using Whisper."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper",
    "click",
    "torch",
    "torchaudio",
    "transformers",
    "accelerate",
    "librosa",
    "reazonspeech-k2-asr @ git+https://github.com/reazon-research/ReazonSpeech.git#subdirectory=pkg/k2-asr"
]

[project.optional-dependencies]
test = [
    "pytest",
]

[project.scripts]
voice-to-text = "src.main:main"
